@@ -8,8 +8,8 @@
     "@testing-library/user-event": "^12.1.10",
     "async-mutex": "^0.3.1",
     "crc": "^3.8.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3",
     "tasktimer": "^3.0.0",
     "typescript": "^4.3.4",
@@ -40,9 +40,9 @@
     ]
   },
   "devDependencies": {
-    "@types/crc": "^3.4.0",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.8",
+    "@types/crc": "^3.8.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/w3c-web-serial": "^1.0.1"
   }
 }
