@@ -1,11 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import {createRoot} from "react-dom/client";
 
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container!);
+
+root.render(
     <React.StrictMode>
         <App />
     </React.StrictMode>,
-document.getElementById('root')
+
 );
